build(deps): adopt published Prompty 2.0.0-beta.4 crates
Retarget the four prompty crates (prompty, prompty-openai, prompty-foundry, prompty-anthropic) off the microsoft/prompty git rev pin and onto the immutable crates.io 2.0.0-beta.4 release (byte-identical tree to main rev 01ef615f). This release carries the unified turn engine, cross-runtime model discovery, and the shared wire::enable_streaming Foundry streaming-usage fix validated live (nonzero total_tokens). Update the single-pin guard test to assert the semver version, and keep the signed usize<->i32 boundary casts in prompty_runner.


Copilot-Session: 3a48a465-fac3-43c9-81a5-d08998029130

+++ src-tauri/src/engine/agent/prompty_runner.rs
@@ -749,9 +749,9 @@ impl HostPolicyPort for CutReadyHostPolicy {
             messages
                 .iter()
                 .take_while(|message| matches!(message.role, Role::System | Role::Developer))
-                .count()
+                .count() as i32
         } else {
-            request.stable_prefix_messages.min(original_len)
+            request.stable_prefix_messages.min(original_len as i32)
         };
 
         Ok(HostPolicyResult {
